net: Fix ttl range
Updated code to correct valid ttl (Time To Live) range.

+++ src/kyron/src/mio/net/tcp_listener.rs
@@ -59,8 +59,8 @@ impl<T: IoSelector> TcpListenerBridge<T> {
         self.inner.as_inner().ttl()
     }
 
-    pub fn set_ttl(&self, ttl: u32) -> IoResult<()> {
-        self.inner.as_inner().set_ttl(ttl)
+    pub fn set_ttl(&self, ttl: u8) -> IoResult<()> {
+        self.inner.as_inner().set_ttl(ttl as u32)
     }
 }
