+++ src/kyron/src/mio/net/tcp_stream.rs
@@ -73,8 +73,8 @@ impl<T: IoSelector> TcpStreamBridge<T> {
         self.inner.as_inner().ttl()
     }
 
-    pub fn set_ttl(&self, ttl: u32) -> IoResult<()> {
-        self.inner.as_inner().set_ttl(ttl)
+    pub fn set_ttl(&self, ttl: u8) -> IoResult<()> {
+        self.inner.as_inner().set_ttl(ttl as u32)
     }
 }
